的の残段数のスクリプトの設置場所を変更した
