+++ Assets/Scripts/Widget/Game/RemaingTargetCounter/IRemainingTargetCounterTextModel.cs
@@ -1,0 +1,17 @@
+using UniRx;
+
+/// <summary>
+/// 的の残段数を管理する
+/// </summary>
+public interface IRemainingTargetCounterTextModel
+{
+    /// <summary>
+    /// 的の残段数
+    /// </summary>
+    public IReactiveProperty<int> RemainingTargetCount { get; }
+   
+    /// <summary>
+    /// 的の残存を減らす
+    /// </summary>
+    public void Decrement();
+}
