+++ Assets/Scripts/Widget/Game/RemaingTargetCounter/RemainingTargetCounterTextModel.cs
@@ -1,24 +1,26 @@
 using UniRx;
 
+/// <summary>
+/// 的の残存のデータを管理する
+/// </summary>
 public class RemainingTargetCounterTextModel : IRemainingTargetCounterTextModel
 {
     /// <summary>
-    /// 
+    /// 的の残存
     /// </summary>
-    public IReactiveProperty<int> RemainingTargetCountProp => _remainingTargetCountProp;
-
+    public IReactiveProperty<int> RemainingTargetCount => _remainingTargetCountProp;
     private IntReactiveProperty _remainingTargetCountProp;
     
     /// <summary>
     /// コンストラクタ
     /// </summary>
-    public RemainingTargetCounterTextModel(int remainingTargetCount)
+    public RemainingTargetCounterTextModel(int targetCount=0)
     {
-        _remainingTargetCountProp = new IntReactiveProperty(remainingTargetCount);
+        _remainingTargetCountProp = new IntReactiveProperty(targetCount);
     }
 
     /// <summary>
-    /// 
+    /// 的の残存を減らす
     /// </summary>
     public void Decrement()
     {
