--- Assets/Scripts/Widget/RemaingTargetCounter/IRemainingTargetCounterTextModel.cs
@@ -1,14 +1,0 @@
-using UniRx;
-
-public interface IRemainingTargetCounterTextModel
-{
-    /// <summary>
-    /// 
-    /// </summary>
-    public IReactiveProperty<int> RemainingTargetCountProp { get; }
-   
-    /// <summary>
-    /// 
-    /// </summary>
-    public void Decrement();
-}
